+++ src/content/style/customBadge.ts
@@ -1,0 +1,97 @@
+export const bestsellerBadge = `
+<div
+  class="a-section a-spacing-none aok-relative puis-status-badge-container s-list-status-badge-container"
+>
+  <div class="a-section a-spacing-none s-badge-spacing">
+    <span
+      data-component-type="s-status-badge-component ecm-badge"
+      class="rush-component"
+      data-component-props='{"asin":"XXXXXXXXXX","badgeType":"ecm-best-seller"}'
+      data-version-id="v2pj4k4hggyykc2ct0d78isz8tt"
+      data-render-id="r36y0io2s2p8sw2dip9doet7xzp"
+      data-component-id="14"
+      ><div class="a-row a-badge-region">
+        <span
+          id="XXXXXXXXXX-best-seller"
+          class="a-badge"
+          role="group"
+          aria-labelledby="XXXXXXXXXX-best-seller-label XXXXXXXXXX-best-seller-supplementary"
+          data-a-badge-supplementary-position="right"
+          tabindex="0"
+          data-a-badge-type="status"
+          ><span
+            id="XXXXXXXXXX-best-seller-label"
+            class="a-badge-label"
+            data-a-badge-color="sx-orange"
+            aria-hidden="true"
+            ><span class="a-badge-label-inner a-text-ellipsis"
+              ><span class="a-badge-text" data-a-badge-color="sx-cloud"
+                >Best Seller</span
+              ></span
+            ></span
+          ></span
+        >
+      </div></span
+    >
+  </div>
+</div>
+`
+
+export const amazonsChoiceBadge = `
+<div
+  class="a-section a-spacing-none aok-relative puis-status-badge-container s-list-status-badge-container"
+>
+  <div class="a-section a-spacing-none s-badge-spacing">
+    <span
+      data-component-type="s-status-badge-component ecm-badge"
+      class="rush-component"
+      data-component-props='{"asin":"XXXXXXXXXX","badgeType":"ecm-amazons-choice"}'
+      data-version-id="v2pj4k4hggyykc2ct0d78isz8tt"
+      data-render-id="r36y0io2s2p8sw2dip9doet7xzp"
+      data-component-id="14"
+      ><div class="a-row a-badge-region">
+        <span
+          id="XXXXXXXXXX-amazons-choice"
+          class="a-badge"
+          role="group"
+          aria-labelledby="XXXXXXXXXX-amazons-choice-label XXXXXXXXXX-amazons-choice-supplementary"
+          data-a-badge-supplementary-position="right"
+          tabindex="0"
+          data-a-badge-type="status"
+          ><span
+            id="XXXXXXXXXX-amazons-choice-label"
+            class="a-badge-label"
+            data-a-badge-color="sx-gulfstream"
+            aria-hidden="true"
+            ><span class="a-badge-label-inner a-text-ellipsis"
+              >
+            <span class="a-badge-text" data-a-badge-color="sx-cloud">Amazon's </span>
+            <span class="a-badge-text" data-a-badge-color="ac-orange">Choice</span>
+              </span
+            ></span
+          ></span
+        >
+      </div></span
+    >
+  </div>
+</div>
+`
+
+export const amazonsChoiceBadgePDP = `
+<div id="acBadge_feature_div" class="celwidget" data-feature-name="acBadge" data-csa-c-type="widget"
+    data-csa-c-content-id="acBadge" data-csa-c-slot-id="acBadge_feature_div" data-csa-c-asin="XXXXXXXXXX"
+    data-csa-c-is-in-initial-active-row="false" data-csa-c-id="532kea-g8j25n-971l2j-k2p4k7"
+    data-cel-widget="acBadge_feature_div">
+    <div class="ac-badge-wrapper" cssclass="aok-float-left">
+        <span class="a-declarative" data-action="a-popover" data-csa-c-type="widget"
+            data-csa-c-func-deps="aui-da-a-popover"
+            data-a-popover="{&quot;name&quot;:&quot;amazons-choice-popover&quot;,&quot;position&quot;:&quot;triggerTop&quot;}"
+            data-csa-c-id="375w0x-xtguck-n80q4l-vgy24r"> <span class="a-size-small aok-float-left ac-badge-rectangle">
+                <span class="ac-badge-text-primary ac-white">Amazon's </span> <span
+                    class="ac-badge-text-secondary ac-orange">Choice</span> </span> <span
+                class="aok-float-left ac-badge-triangle"></span> </span> <span class="ac-for-text">
+            </span>
+    </div>
+    <script type="a-state" data-a-state="{&quot;key&quot;:&quot;acState&quot;}">{"acAsin":"XXXXXXXXXX"}</script>
+</div>
+`